Fusionné les deux controlleurs qui travaillent sur les mouvements de dossiers en masse (consultation et traitement) en un seul en passant à Spring 3
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@68824 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/core/src/main/java/ch/vd/uniregctb/mouvement/MouvementDossierDAO.java
@@ -41,7 +41,7 @@ public interface MouvementDossierDAO extends GenericDAO<MouvementDossier, Long> 
 	 * @param criteria
 	 * @return
 	 */
-	long count(MouvementDossierCriteria criteria);
+	int count(MouvementDossierCriteria criteria);
 
 	/**
 	 * Recherche de plusieurs mouvements de dossier par ID en même temps
